+++ game/ecs/sys/sysRectCollision.go
@@ -11,7 +11,7 @@ type SystemRectCollision struct {
 
 }
 
-func (sr *SystemRectCollision) Update(world inter.WorldSpace, entities *[]ec.Entity, delta float64) {
+func (sr *SystemRectCollision) Update(world inter.WorldSpace, entities *[]ec.Entity, delta float32) {
 	for i := range *entities {
 		if (*entities)[i].HasComponent(ec.C_SPATIAL) && (*entities)[i].HasComponent(ec.C_CLICKER) {
 			//Get an entity with spatial
